[SW-73] refactor: interceptor 파일 분리
